+ Automatic Abonnement http://teamopenuss.uni-muenster.de/jira/browse/USS-884 # http://teamopenuss.uni-muenster.de/jira/browse/USS-880
git-svn-id: https://svn.code.sf.net/p/openuss/code/branches/openuss-bachelor-ws0708@2856 60046477-0c10-0410-bf4b-e552a4ff1768

+++ plexus/plexus-core/src/main/java/org/openuss/discussion/DiscussionServiceImpl.java
@@ -316,9 +316,12 @@ public class DiscussionServiceImpl extends DiscussionServiceBase {
 	protected void handleRemoveForumWatch(ForumInfo forum) throws Exception {
 		Validate.notNull(forum);
 		Validate.notNull(forum.getId());
+		
 		Forum f = getForumDao().load(forum.getId());
 		ForumWatch fw = getForumWatchDao().findByUserAndForum(getSecurityService().getCurrentUser(), f);
-		getForumWatchDao().remove(fw);
+		if(fw != null) {
+			getForumWatchDao().remove(fw);
+		}
 	}
 
 	/**
